dts: arm64: renesas: r8a779g0: enable scmi reset domains
Enable SCMI reset domains support.

This patch also replaces all reset_id values in DT with defines which are
getting the same values if SCMI=off or sequential numbers in case SCMI=on.

The current Reset domain identification approach is not compatible with
SCMI which expected Reset domains to be enumerated sequentially.

Current:
 resets = <&cpg 915>; where 915 = 9 * 100 + 15, 9 is register number and 15
is bit number.

After this patch:
no_scmi:
 resets = <&cpg R8A779G0_RESET_915>; R8A779G0_RESET_915==915
scmi:
 resets = <&scmi_reset R8A779G0_RESET_915>; R8A779G0_RESET_915==114

Signed-off-by: Grygorii Strashko <grygorii_strashko@epam.com>