dts: arm64: renesas: r8a779g0: enable scmi clocks domains
Enable SCMI Clocks  support.

This patch also replaces all clock_id values in DT with defines which are
getting the same values if SCMI=off or sequential numbers in case SCMI=on.

The current Clocks identification approach is not compatible with SCMI
which expected Clocks to be enumerated sequentially. Only Core clocks
enumerated in compatible way, but module clocks - not.

Current:
 clocks = <&cpg CLK_MOD 915>; where 915 = 9 * 100 + 15, 9 is register
number and 15 is bit number.

SCMI Clocks enumerated as:
 - core clks 0..N
 - mod clks N+1..M

After this patch:
no_scmi:
  clocks = <&cpg CLK_MOD R8A779G0_CLK_MOD_514>,
  //R8A779G0_CLK_MOD_514==514
           <&cpg CLK_CORE R8A779G0_CLK_SASYNCPERD1>,
           <&scif_clk>;
scmi:
  clocks = <&scmi_clk R8A779G0_CLK_MOD_514>,
  //R8A779G0_CLK_MOD_514==(R8A779G0_CLK_MOD_FIRST + 32)
           <&scmi_core_clk R8A779G0_CLK_SASYNCPERD1>,
           <&scif_clk>;

Signed-off-by: Grygorii Strashko <grygorii_strashko@epam.com>